requirements.txt
discord.py>=2.3.0
instaloader>=4.10
python-dotenv>=1.0.0
Flask~=3.1.3
